pyproject.toml
[tool.poetry]
name = "bapi"
version = "2.2.0"
description = "BeeStation public api"
authors = ["Qwertyquerty", "Crossedfall"]
license = "APACHE"
readme = "readme.md"
repository = "https://github.com/BeeStation/bapi"
packages = [{include = "bapi", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
cachetools = "~6.1.0"
elastic-apm = {version = "~6.23.0", extras = ["flask"]}
requests = "~2.32.4"
mysqlclient = "~2.2.7"
patreon = {git = "https://github.com/Patreon/patreon-python.git", rev = "80c83f0"}
PyYAML = "~6.0.2"
Flask = "~3.1.1"
Flask-Cors = "~6.0.1"
Flask-SQLAlchemy = "~3.1.1"
uWSGI = "~2.0.30"
werkzeug = "~3.1.3"
marshmallow-sqlalchemy = "~1.4.2"
setuptools = "~80.9.0"
discord-oauth2-py = "~1.2.2"
flask-smorest = "~0.46.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
black = "^25.1.0"
pre-commit = "^4.2.0"
codespell = "^2.4.1"
mypy = "^1.16.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.codespell]
count = ""
skip = ".git,.venv,*.lock,.mypy_cache"
builtin = "clear,rare,code,en-GB_to_en-US"
ignore-words-list = "od,ro"
quiet = 3

[tool.mypy]
mypy_path = "bapi"
check_untyped_defs = true
disallow_any_generics = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
no_implicit_reexport = true

[[tool.mypy.overrides]]
module = "bapi.db"
disable_error_code = "name-defined"
